Refactor(Solution): Downgrade .NET Core 3.1 to 3.0 Fetch(MatrixCalculator.cs): Pow method works with negative degrees. Refactor(GUI): Minimize window and handle for set size matrix elem. Test(PowTest.cs ComplexTest.cs): Testing negative Pow method and complex matrix parser expression. Fetch(License.WTFPL README.md): Add License and README.

+++ MatrixCalculator/MatrixExpressionParser.cs
@@ -167,7 +167,7 @@ namespace MatrixCalculator
                             return new Result(MatrixMath.Transpose(accum), cur.Expression);
                         }
 
-                        return new Result(MatrixMath.Pow(accum, cur.Accumulator[0, 0]), cur.Expression);
+                        return new Result(MatrixMath.Pow(accum, (short)cur.Accumulator[0, 0]), cur.Expression);
                 }
             }
